INS-30: Move storage types into storage dir.
