+++ ledger/storage/record/result.go
