+++ ledger/storage/record/serialize.go
