+++ ledger/storage/record/serialize_test.go
